requirements.txt
netmiko
pyyaml
colorama